+++ lib/flutter_ocr_sdk.dart
@@ -1,6 +1,7 @@
 import 'dart:typed_data';
 
 import 'flutter_ocr_sdk_platform_interface.dart';
+import 'mrz_line.dart';
 
 class FlutterOcrSdk {
   Future<String?> getPlatformVersion() {
@@ -12,25 +13,15 @@ class FlutterOcrSdk {
     return FlutterOcrSdkPlatform.instance.init(path, key);
   }
 
-  Future<String?> recognizeByBuffer(
+  Future<List<List<MrzLine>>?> recognizeByBuffer(
       Uint8List bytes, int width, int height, int stride, int format) {
     return FlutterOcrSdkPlatform.instance
         .recognizeByBuffer(bytes, width, height, stride, format);
   }
 
-  Future<String?> recognizeByFile(String filename) {
+  Future<List<List<MrzLine>>?> recognizeByFile(String filename) {
     return FlutterOcrSdkPlatform.instance.recognizeByFile(filename);
   }
-
-  // Future<int?> loadModelFiles(String name, Uint8List prototxtBuffer,
-  //     Uint8List txtBuffer, Uint8List characterModelBuffer) {
-  //   return FlutterOcrSdkPlatform.instance
-  //       .loadModelFiles(name, prototxtBuffer, txtBuffer, characterModelBuffer);
-  // }
-
-  // Future<int?> loadTemplate(String template) async {
-  //   return FlutterOcrSdkPlatform.instance.loadTemplate(template);
-  // }
 
   Future<int?> loadModel(String modelPath) async {
     return FlutterOcrSdkPlatform.instance.loadModel(modelPath);