+++ lib/flutter_ocr_sdk_platform_interface.dart
@@ -2,6 +2,49 @@ import 'package:flutter/services.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'flutter_ocr_sdk_method_channel.dart';
+import 'mrz_line.dart';
+
+/// Image pixel format.
+enum ImagePixelFormat {
+  /// 0:Black, 1:White
+  IPF_BINARY,
+
+  /// 0:White, 1:Black
+  IPF_BINARYINVERTED,
+
+  /// 8bit gray
+  IPF_GRAYSCALED,
+
+  /// NV21
+  IPF_NV21,
+
+  /// 16bit with RGB channel order stored in memory from high to low address
+  IPF_RGB_565,
+
+  /// 16bit with RGB channel order stored in memory from high to low address
+  IPF_RGB_555,
+
+  /// 24bit with RGB channel order stored in memory from high to low address
+  IPF_RGB_888,
+
+  /// 32bit with ARGB channel order stored in memory from high to low address
+  IPF_ARGB_8888,
+
+  /// 48bit with RGB channel order stored in memory from high to low address
+  IPF_RGB_161616,
+
+  /// 64bit with ARGB channel order stored in memory from high to low address
+  IPF_ARGB_16161616,
+
+  /// 32bit with ABGR channel order stored in memory from high to low address
+  IPF_ABGR_8888,
+
+  /// 64bit with ABGR channel order stored in memory from high to low address
+  IPF_ABGR_16161616,
+
+  /// 24bit with BGR channel order stored in memory from high to low address
+  IPF_BGR_888
+}
 
 abstract class FlutterOcrSdkPlatform extends PlatformInterface {
   /// Constructs a FlutterOcrSdkPlatform.
@@ -32,23 +75,14 @@ abstract class FlutterOcrSdkPlatform extends PlatformInterface {
     throw UnimplementedError('init() has not been implemented.');
   }
 
-  Future<String?> recognizeByBuffer(
+  Future<List<List<MrzLine>>?> recognizeByBuffer(
       Uint8List bytes, int width, int height, int stride, int format) {
     throw UnimplementedError('recognizeByBuffer() has not been implemented.');
   }
 
-  Future<String?> recognizeByFile(String filename) {
+  Future<List<List<MrzLine>>?> recognizeByFile(String filename) {
     throw UnimplementedError('recognizeByFile() has not been implemented.');
   }
-
-  // Future<int?> loadModelFiles(String name, Uint8List prototxtBuffer,
-  //     Uint8List txtBuffer, Uint8List characterModelBuffer) {
-  //   throw UnimplementedError('loadModelFiles() has not been implemented.');
-  // }
-
-  // Future<int?> loadTemplate(String template) async {
-  //   throw UnimplementedError('loadTemplate() has not been implemented.');
-  // }
 
   Future<int?> loadModel(String modelPath) async {
     throw UnimplementedError('loadModel() has not been implemented.');