+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+- Added `MrzLine` class.
+- Added image pixel format enum.
+- Updated recognition APIs to return `List<List<MrzLine>>` instead of `JSON String`.
+
 ## 0.1.0
 
 * Updated Dynamsoft Label Recognizer SDK.